# Conflicts:
#	app/controllers/tour_stops_controller.rb
#	app/controllers/tours_controller.rb
#	app/uploaders/avatar_uploader.rb
#	app/views/tours/index.html.haml

+++ app/controllers/tour_stops_controller.rb
@@ -14,21 +14,18 @@ class TourStopsController < ApplicationController
   end
   def destroy
     @tour_stop = TourStop.find params[:id]
-    paths = Path.where(map_path_to_id: @tour_stop.stop_id)
-    paths.each do |path|
-      path.path_points.destroy_all
-      path.destroy if path.present?
-    end
     paths = Path.where(map_path_from_id: @tour_stop.stop_id)
     paths.each do |path|
       path.path_points.destroy_all
       path.destroy if path.present?
     end
-    # path = @tour_stop.stop_type.classify.constantize.find_by_id(@tour_stop.stop_id).paths.last
-    # path.path_points.destroy_all if path.present?
+
+    path = @tour_stop.stop_type.classify.constantize.find_by_id(@tour_stop.stop_id).paths.last
+    path.path_points.destroy_all if path.present?
+    path.destroy if path.present?
     VisitedStop.where(tour_stop_id: @tour_stop.id).destroy_all
     if @tour_stop.destroy
-      redirect_to community_tours_path(current_community), :notice => "Tour Stop deleted"
+      redirect_to community_tours_path(current_community,floorNo = (params[:floorplate].present? ? params[:floorplate] : nil)), :notice => "Tour Stop deleted"
     else
       redirect_to community_tours_path(current_community)
     end
